requirements.txt
# /requirements.txt

django
djangorestframework
djangorestframework-simplejwt
flask
python-dotenv
python-jose
flask-cors
six